backend/utils/ollama.py
import httpx
import json
import os

from dotenv import load_dotenv

load_dotenv()

OLLAMA_URL = os.getenv("OLLAMA_URL", "http://localhost:11434/api/generate")
OLLAMA_MODEL = os.getenv("OLLAMA_MODEL", "meta-llama/Llama-3-8b") # or mistral, etc.

async def generate_completion(prompt: str) -> str:
    """Make real HTTP call to Ollama."""
    try:
        async with httpx.AsyncClient() as client:
            response = await client.post(
                OLLAMA_URL,
                json={
                    "model": OLLAMA_MODEL,
                    "prompt": prompt,
                    "stream": False,
                },
                timeout=120.0
            )
            response.raise_for_status()
            data = response.json()
            return data.get("response", "")
    except Exception as e:
        print(f"Error calling Ollama: {e}")
        return f"Error: {str(e)}"
